refactor: isolate workspace diff tools
