+++ inc/Abilities/WorkspaceAbilities.php
@@ -716,87 +716,6 @@ class WorkspaceAbilities {
 			);
 
 			wp_register_ability(
-				'datamachine/workspace-diff-summary',
-				array(
-					'label'               => 'Workspace Diff Summary',
-					'description'         => 'Summarize changed files and compact diff metadata for a workspace handle.',
-					'category'            => 'datamachine-code-workspace',
-					'input_schema'        => array(
-						'type'       => 'object',
-						'properties' => array(
-							'name'   => array(
-								'type'        => 'string',
-								'description' => 'Workspace handle: `<repo>` (primary) or `<repo>@<branch-slug>` (worktree).',
-							),
-							'from'   => array(
-								'type'        => 'string',
-								'description' => 'Optional from git ref.',
-							),
-							'to'     => array(
-								'type'        => 'string',
-								'description' => 'Optional to git ref.',
-							),
-							'staged' => array(
-								'type'        => 'boolean',
-								'description' => 'Summarize staged diff instead of working tree diff.',
-							),
-							'path'   => array(
-								'type'        => 'string',
-								'description' => 'Optional relative path filter.',
-							),
-						),
-						'required'   => array( 'name' ),
-					),
-					'output_schema'       => array( 'type' => 'object' ),
-					'execute_callback'    => array( self::class, 'diffSummary' ),
-					'permission_callback' => fn() => PermissionHelper::can_manage(),
-					'meta'                => array( 'show_in_rest' => true ),
-				)
-			);
-
-			wp_register_ability(
-				'datamachine/workspace-diff-validate',
-				array(
-					'label'               => 'Validate Workspace Diff',
-					'description'         => 'Validate workspace diff shape using allowed/denied path patterns and optional test-change requirements.',
-					'category'            => 'datamachine-code-workspace',
-					'input_schema'        => array(
-						'type'       => 'object',
-						'properties' => array(
-							'name'                  => array( 'type' => 'string' ),
-							'from'                  => array( 'type' => 'string' ),
-							'to'                    => array( 'type' => 'string' ),
-							'staged'                => array( 'type' => 'boolean' ),
-							'path'                  => array( 'type' => 'string' ),
-							'allow'                 => array(
-								'type'  => 'array',
-								'items' => array( 'type' => 'string' ),
-							),
-							'deny'                  => array(
-								'type'  => 'array',
-								'items' => array( 'type' => 'string' ),
-							),
-							'include_any'           => array(
-								'type'  => 'array',
-								'items' => array( 'type' => 'string' ),
-							),
-							'include_all'           => array(
-								'type'  => 'array',
-								'items' => array( 'type' => 'string' ),
-							),
-							'require_tests'         => array( 'type' => 'boolean' ),
-							'require_changed_files' => array( 'type' => 'object' ),
-						),
-						'required'   => array( 'name' ),
-					),
-					'output_schema'       => array( 'type' => 'object' ),
-					'execute_callback'    => array( self::class, 'diffValidate' ),
-					'permission_callback' => fn() => PermissionHelper::can_manage(),
-					'meta'                => array( 'show_in_rest' => true ),
-				)
-			);
-
-			wp_register_ability(
 				'datamachine/workspace-git-pull',
 				array(
 					'label'               => 'Workspace Git Pull',
@@ -2806,41 +2725,5 @@ class WorkspaceAbilities {
 			! empty( $input['staged'] ),
 			$input['path'] ?? null
 		);
-	}
-
-	/**
-	 * Summarize git diff metadata for a workspace repository.
-	 *
-	 * @param array $input Input parameters.
-	 * @return array|\WP_Error
-	 */
-	public static function diffSummary( array $input ): array|\WP_Error {
-		if ( RemoteWorkspaceBackend::should_handle() ) {
-			return new \WP_Error( 'workspace_diff_summary_remote_unsupported', 'Workspace diff summary currently supports local workspaces only.', array( 'status' => 501 ) );
-		}
-
-		$workspace = new Workspace();
-		return $workspace->diff_summary(
-			$input['name'] ?? '',
-			$input['from'] ?? null,
-			$input['to'] ?? null,
-			! empty( $input['staged'] ),
-			$input['path'] ?? null
-		);
-	}
-
-	/**
-	 * Validate git diff metadata for a workspace repository.
-	 *
-	 * @param array $input Input parameters.
-	 * @return array|\WP_Error
-	 */
-	public static function diffValidate( array $input ): array|\WP_Error {
-		if ( RemoteWorkspaceBackend::should_handle() ) {
-			return new \WP_Error( 'workspace_diff_validate_remote_unsupported', 'Workspace diff validation currently supports local workspaces only.', array( 'status' => 501 ) );
-		}
-
-		$workspace = new Workspace();
-		return $workspace->diff_validate( $input['name'] ?? '', $input );
 	}
 }
